+++ lib/main.dart
@@ -1,6 +1,5 @@
-import 'package:alfa_soyzen/presentation/splash_screen.dart';
+import 'package:alfa_soyzen/presentation/splashscreen/splash_screen.dart';
 import 'package:flutter/material.dart';
-import 'package:alfa_soyzen/presentation/homescreen.dart';
 
 void main() => runApp(const MyApp());
 
@@ -14,10 +13,7 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      home: HomeScreen(),
-      routes: {
-        '/notification': (context) => const SplashScreen(),
-      },
+      home: const SplashScreen(),
     );
   }
 }
