+++ lib/presentation/onboarding/onboarding_screen_1.dart
@@ -1,94 +1,84 @@
-import 'package:alfa_soyzen/presentation/onboarding_screen_1.dart';
-import 'package:alfa_soyzen/presentation/onboarding_screen_3.dart';
-import 'package:alfa_soyzen/presentation/verification_screen.dart';
+import 'package:alfa_soyzen/presentation/onboarding/onboarding_screen_2.dart';
+import 'package:alfa_soyzen/presentation/onboarding/onboarding_screen_3.dart';
+import 'package:alfa_soyzen/presentation/auth/verification_screen.dart';
 import 'package:flutter/material.dart';
 
-void main() => runApp(const OnBoardingScreen2());
+void main() => runApp(const OnBoardingScreen1());
 
-class OnBoardingScreen2 extends StatelessWidget {
-  const OnBoardingScreen2({super.key});
+class OnBoardingScreen1 extends StatelessWidget {
+  const OnBoardingScreen1({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return const Scaffold(
+    return Scaffold(
         body: Center(
             child: Column(
       children: [
-        SizedBox(height: 90),
-        YogaImage(),
-        SizedBox(height: 53),
-        YogaDaily(),
-        SizedBox(height: 90),
-        BarraNavegacion(),
+        const SizedBox(height: 50),
+        yogaImage(),
+        const SizedBox(height: 20),
+        yogaDaily(),
+        const SizedBox(height: 90),
+        const BarraNavegacion(),
       ],
     )));
   }
 }
 
-class YogaImage extends StatelessWidget {
-  const YogaImage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      height: 231.01,
-      width: 329.0,
-      child: Image.asset('assets/images/yoga2.png'),
-    );
-  }
+Widget yogaImage() {
+  return SizedBox(
+    height: 304.0,
+    width: 328.0,
+    child: Image.asset('assets/images/yoga1.png'),
+  );
 }
 
-class YogaDaily extends StatelessWidget {
-  const YogaDaily({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Column(
-      children: [
-        SizedBox(
-          width: 109,
-          height: 28,
-          child: Text('Meditation',
-              textAlign: TextAlign.center,
-              style: TextStyle(
-                color: Color(0xFF4F14A0),
-                fontWeight: FontWeight.w400,
-                fontSize: 22,
-                fontFamily: 'PTSans',
-              )),
-        ),
-        SizedBox(height: 10),
-        SizedBox(
-            width: 233,
-            height: 49,
-            child: Text(
-              'Yoga Classes',
-              textAlign: TextAlign.center,
-              style: TextStyle(
-                  color: Color(0xFF222222),
-                  fontWeight: FontWeight.w700,
-                  fontSize: 38,
-                  fontFamily: 'PTSans'),
-            )),
-        SizedBox(height: 20),
-        SizedBox(
-          width: 260,
-          height: 45,
-          child: Text(
-            'Meditation is the key to Productivity. \nHappiness & Longevity',
+Widget yogaDaily() {
+  return const Column(
+    children: [
+      SizedBox(
+        width: 55,
+        height: 28,
+        child: Text('Yoga',
             textAlign: TextAlign.center,
             style: TextStyle(
-                color: Color(0xFF677294),
-                fontWeight: FontWeight.w400,
-                fontSize: 15,
+              color: Color(0xFF4F14A0),
+              fontWeight: FontWeight.w400,
+              fontSize: 22,
+              fontFamily: 'PTSans',
+            )),
+      ),
+      SizedBox(height: 10),
+      SizedBox(
+          width: 190,
+          height: 49,
+          child: Text(
+            'Daily Yoga',
+            textAlign: TextAlign.center,
+            style: TextStyle(
+                color: Color(0xFF222222),
+                fontWeight: FontWeight.w700,
+                fontSize: 38,
                 fontFamily: 'PTSans'),
-          ),
+          )),
+      SizedBox(height: 20),
+      SizedBox(
+        width: 290,
+        height: 45,
+        child: Text(
+          'Do your practice of physical exercise and \n relaxation make helthy',
+          textAlign: TextAlign.center,
+          style: TextStyle(
+              color: Color(0xFF677294),
+              fontWeight: FontWeight.w400,
+              fontSize: 15,
+              fontFamily: 'PTSans'),
         ),
-        SizedBox(height: 35),
-        Buttom(),
-      ],
-    );
-  }
+      ),
+      SizedBox(height: 35),
+      Buttom(),
+    ],
+  );
 }
 
 class Buttom extends StatelessWidget {
@@ -101,7 +91,7 @@ class Buttom extends StatelessWidget {
       child: ElevatedButton(
         onPressed: () {
           Navigator.of(context).pushReplacement(MaterialPageRoute(
-            builder: (_) => const OnBoardingScreen3(),
+            builder: (_) => const OnBoardingScreen2(),
           ));
         },
         child: const Row(
@@ -165,16 +155,16 @@ class BarraNavegacion extends StatelessWidget {
                 mainAxisAlignment: MainAxisAlignment.end,
                 children: <Widget>[
                   IconButton(
+                    icon: const Icon(Icons.radio_button_checked),
+                    onPressed: () {},
+                  ),
+                  IconButton(
                     icon: const Icon(Icons.radio_button_unchecked),
                     onPressed: () {
                       Navigator.of(context).pushReplacement(MaterialPageRoute(
-                        builder: (_) => const OnBoardingScreen1(),
+                        builder: (_) => const OnBoardingScreen2(),
                       ));
                     },
-                  ),
-                  IconButton(
-                    icon: const Icon(Icons.radio_button_checked),
-                    onPressed: () {},
                   ),
                   IconButton(
                     icon: const Icon(Icons.radio_button_unchecked),
